requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
pandas>=1.1.4
seaborn>=0.11.0
scikit-learn==1.1.1
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
tqdm>=4.41.0

# tests/linter
black
coverage
flake8
pytest
yapf

# utilities
six>=1.14
joblib
memoized-property
termcolor

# interface
streamlit
snowflake-connector-python
streamlit-modal
